ml_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-multipart==0.0.9

# NLP / Embeddings
sentence-transformers==2.7.0
transformers==4.41.2
torch==2.3.0

# Computer Vision (CLIP)
open-clip-torch==2.24.0
Pillow==10.3.0
requests==2.31.0

# ML / Clustering
scikit-learn==1.5.0
numpy==1.26.4
scipy==1.13.0

# Geo
geopy==2.4.1

# HTTP client for downstream calls
httpx==0.27.0

# Colab / ngrok
pyngrok==7.1.6
